feat: Update event logos and speakers; refactor sponsors section for IWD 2026
