+++ src/components/pages/GDGBHMeet/SpeakersGDGBHMeet2/index.tsx
@@ -33,28 +33,22 @@ export default function SpeakersGDGBHMeet2() {
             <SectionTitle highlight="Palestrantes" />
             <div className="mb-16"></div>
             <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-                {speakers.length > 0 ? (
-                    <div className="flex flex-wrap justify-center gap-6">
-                        {speakers.map((speaker, idx) => (
-                            <div key={idx} className="w-full sm:w-[280px]">
-                                <ProfileCard
-                                    photoUrl={speaker.photoUrl}
-                                    name={speaker.name}
-                                    role={speaker.role}
-                                    description={speaker.description}
-                                    links={speaker.links}
-                                    speakerInfo={speaker.speakerInfo}
-                                    talkTitle={speaker.talkTitle}
-                                    talkInfo={speaker.talkInfo}
-                                />
-                            </div>
-                        ))}
-                    </div>
-                ) : (
-                    <div className="text-center text-gray-600">
-                        <p className="text-xl font-medium">Em breve divulgaremos os palestrantes desta edição!</p>
-                    </div>
-                )}
+                <div className="flex flex-wrap justify-center gap-6">
+                    {speakers.map((speaker, idx) => (
+                        <div key={idx} className="w-full sm:w-[280px]">
+                            <ProfileCard
+                                photoUrl={speaker.photoUrl}
+                                name={speaker.name}
+                                role={speaker.role}
+                                description={speaker.description}
+                                links={speaker.links}
+                                speakerInfo={speaker.speakerInfo}
+                                talkTitle={speaker.talkTitle}
+                                talkInfo={speaker.talkInfo}
+                            />
+                        </div>
+                    ))}
+                </div>
             </div>
         </section>
     );
