+++ src/components/pages/IWD/SpeakersIWD2026/index.tsx
@@ -12,6 +12,14 @@ export default function SpeakersIWD2026() {
             speakerInfo: 'Com mais de 10 anos de experiência em RH em grandes empresas nacionais e multinacionais — como Souza Cruz, Midea Carrier, Exxon Mobil, Disys, Arezzo e SouCloud —, Luciana é hoje Consultora de RH e influenciadora digital, conhecida como o “RH mais Sincero do Brasil”. Ela é também criadora do polêmico Manual de Sobrevivência Corporativo: Curso de Puxa Saco, que traz uma visão crítica e bem-humorada sobre o mundo corporativo.',
             talkTitle: 'Entrevista de Emprego sem Frescura',
             talkInfo: 'Nesta palestra, Luciana vai abordar os principais pilares de uma entrevista de emprego:\n✔️ Preparação e análise da vaga\n✔️ Como responder às principais perguntas\n✔️ Os erros mais comuns cometidos pelos candidatos\n\nTudo de forma descontraída, direta e real, mostrando como as coisas realmente funcionam no mercado de trabalho.'
+        },
+        {
+            photoUrl: new URL('/src/assets/palestrantes/Manoela_Werneck.jpeg', import.meta.url).href,
+            name: 'Manoela Werneck',
+            role: 'Tech Lead',
+            description: 'SYDLE',
+            links: { linkedin: 'https://www.linkedin.com/in/manoela-werneck/' },
+            speakerInfo: 'Manoela iniciou sua trajetória na SYDLE há cinco anos, atuando como desenvolvedora full stack em projetos estratégicos nas áreas de mercado livre de energia, billing e e-commerce. Ao longo da sua caminhada, também compartilhou sua experiência em palestras e eventos, com foco em resolução de problemas e liderança técnica.\nAtualmente, lidera um time de desenvolvimento de BPMS, sendo responsável desde a concepção da arquitetura até a implementação final das soluções.\nUma trajetória que une profundidade técnica, visão estratégica e liderança na prática.'
         }
     ];
 
@@ -20,31 +28,22 @@ export default function SpeakersIWD2026() {
             <SectionTitle highlight="Palestrantes" />
             <div className="mb-16"></div>
             <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-                {speakers.length > 0 ? (
-                    <div className="flex flex-wrap justify-center gap-6">
-                        {speakers.map((speaker, idx) => (
-                            <div key={idx} className="w-full sm:w-[280px]">
-                                <ProfileCard
-                                    photoUrl={speaker.photoUrl}
-                                    name={speaker.name}
-                                    role={speaker.role}
-                                    description={speaker.description}
-                                    links={speaker.links}
-                                    speakerInfo={speaker.speakerInfo}
-                                    talkTitle={speaker.talkTitle}
-                                    talkInfo={speaker.talkInfo}
-                                />
-                            </div>
-                        ))}
-                    </div>
-                ) : (
-                    <div className="text-center text-gray-600">
-                        <p className="text-xl font-medium">Em breve divulgaremos as palestrantes desta edição especial!</p>
-                        <p className="text-base mt-4 text-gray-500">
-                            Estamos preparando uma programação incrível com mulheres inspiradoras da tecnologia.
-                        </p>
-                    </div>
-                )}
+                <div className="flex flex-wrap justify-center gap-6">
+                    {speakers.map((speaker, idx) => (
+                        <div key={idx} className="w-full sm:w-[280px]">
+                            <ProfileCard
+                                photoUrl={speaker.photoUrl}
+                                name={speaker.name}
+                                role={speaker.role}
+                                description={speaker.description}
+                                links={speaker.links}
+                                speakerInfo={speaker.speakerInfo}
+                                talkTitle={speaker.talkTitle}
+                                talkInfo={speaker.talkInfo}
+                            />
+                        </div>
+                    ))}
+                </div>
             </div>
         </section>
     );
